refactor: seperate view and controller in Manage
+ VendingMachineManage.js
Controller

+ VendingMachineManageTemplate.js
View
